+++ src/test/java/nextstep/subway/AcceptanceTest.java
@@ -1,18 +1,11 @@
 package nextstep.subway;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import java.util.Map;
 import nextstep.subway.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AcceptanceTest {
@@ -26,19 +19,5 @@ public class AcceptanceTest {
     public void setUp() {
         RestAssured.port = port;
         databaseCleanup.execute();
-    }
-
-    public static <T> ExtractableResponse<Response> get(String path, Map<String, T> params) {
-        return RestAssured
-                .given().log().all()
-                .queryParams(params)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get(path)
-                .then().log().all()
-                .extract();
-    }
-
-    public static void 응답결과_확인(ExtractableResponse<Response> response, HttpStatus httpStatus) {
-        assertThat(response.statusCode()).isEqualTo(httpStatus.value());
     }
 }
